OCR-main/requirements.txt


torch
torchvision
numpy


Pillow
lmdb


streamlit